feat(verify): only promote if verifications succeed (#1705)
* feat(verify): only promote on verif success

This is work towards supporting verification.

Problem

We don't want to promote a version to the next environment if there are verifications that haven't succeeded yet.

Proposed solution

Extend the behavior of the DependsOnConstraintEvaluator to enforce that all verifications have passed in the required environment before allowing promotion.

The DependsOnConstraintEvaluator seemed like a natural place to put this logic.

Future work

Provide feedback to the end user about the status of verifications. This
code doesn't provide any user/operator feedback about the status of
verifications. That will likely live elsewhere in the codebase.

* fix(pr): zap whitespace

* fix(pr): add getStates method

* fix(pr): change method sigs, add logging

Change the method signatures for the verification repository

Add logging around verification queries

* fix(pr): refactor logging to use also

* fix(pr): refactor - pull out method

* fix(pr): remove a ridiculous inefficiency

Call the database only once

* fix(pr): add missing mock