fix(aws/titus cache): fix cluster and application cache eviction (#5213)
Clusters and Applications are global entities and can not be
authoritatively cached by a regionally scoped caching agent.

This moves them to INFORMATIVE and adds a new ClusterCleanupAgent
to handle eviction of clusters with no serverGroups.

Also includes changes in the Application and Cluster providers to
account for inconsistent reads (particularly when doing
titus streaming updates) by looking for the presence of cluster
or server groups rather than assuming the application -> cluster
relationship is valid.