chore(test): rename test label so 'failed' doesn't appear (#1809)
Problem

When tests fail on GitHub Actions, I search for the string `" failed"`
to jump to the failing test lines.

When I do that, the first search hit is this test, which is successful,
but has the text `"failed"` in the name.

```
BakeryLifecycleMonitorTests publishes failed event and marks failure PASSED
```

Implemented solution

Change the text from `failed` to `fail(ed)` so it doesn't match a `
failed` setring search.