feat(interlink): support for restarting stages of foreign executions (#4066)
* feat(interlink): support for restarting stages of foreign executions

Without this PR, calling the restart endpoint causes orca to push a RestartStage message on the queue. But for foreign executions, this doesn't do anything.

With this PR, we follow the model of delegating to a CompoundExecutionOperator instead, which handles local vs foreign executions and queue vs execution repository routing. If the execution is foreign, it will rely on the execution repository to emit the appropriate interlink message.

* chore(interlink): remove source of spurious NPEs