fix(db): handle verif query when artifact ref and env name are the same (#1836)
## Problem

If a delivery config contains an artifact reference and environment that both happen to have the same name (e.g., an artifact reference named `"dev"` and an environment named `"dev"`), then the [`SqlVerificationRepository.getStatesBatch`][1] method throws a SQLSyntaxErrorException.

The problem is described in detail in issue #1821.

## Implemented solution

Add field aliases to prevent the name collision.

Fixes #1821

[1]: https://github.com/spinnaker/keel/blob/d87062779779c383685a2f495c1ce2dc719dbf6f/keel-sql/src/main/kotlin/com/netflix/spinnaker/keel/sql/SqlVerificationRepository.kt#L183-L241