fix(db): add read lock to avoid phantom read (#1838)
## Problem

We've seen cases where two nodes check the same environments for verification within about the same second, even though verification checks are only supposed to have for an environment once a minute.

```
02/23/2021 12:46:35.613 -0800
02/23/2021 12:46:35.869 -0800
```

### Hypothesis: phantom reads

I belive the problem is related to the database consistency level used in the [SqlVerificationRepository.nextEnvironmentsForVerification][1] check. This is a databasde anomaly known as *phantom reads*.

See #1832 for details.

## Implemented solution: locking reads

This PR adds a [`LOCK IN SHARE MODE`][2] directive to the relevant select statement to ensure that phantom reads can't happen.

## Concerns: performance impact

I don't know how this lock will impact database performance. As a precaution, I put in a feature toggle so we can turn it off quickly: `keel.verifications.db.lock.reads.enabled`

I set it to be on by default, so we can see what effect it has as soon as it rolls out.

[1]: https://github.com/spinnaker/keel/blob/d87062779779c383685a2f495c1ce2dc719dbf6f/keel-sql/src/main/kotlin/com/netflix/spinnaker/keel/sql/SqlVerificationRepository.kt#L59-L127
[2]: https://dev.mysql.com/doc/refman/5.7/en/innodb-locking-reads.html