perf(events): create execution events with a ref to the execution (#4076)
* perf(events): create execution events with a ref to the execution

This avoid loading things for the repository twice, e.g.:
- on a StartTask message, the handler loads the entire execution
- it then emits a TaskStarted event, which will load the entire execution again
- ... and so on for TaskComplete, StageStarted, StageComplete, etc.

On simple executions, it decreases the amount of executions fetched by 40%, which should translate to less load on the database and lower memory pressure on orca.

* fix(test): reset state in StartExecutionHandlerTest

The test reuses state between tests, and StartExecutionHandler now mutates the status of the execution so we need to reset it.