feat(admin): api to force skip and fail (#1912)
## Context

In the process of implementing checks to prevent concurrent verifications running in the same environment, or verifications running at the same time as a deployment in an environment.

## Problem

The enforcement checks can block verifications or deployments if a verification gets stuck in a PENDING state or a deployment gets stuck in a DEPLOYING state.

## Implemented solution

Provide admin APIs that allow an operator to force:
* an environment artifact version to SKIPPED state
* a verification into OVERRIDE_FAIL state