chore(rxjs): switch back to bundling rxjs 5.x instead of migrating to rxjs (#583)
There are breaking changes in rxjs6, which deck is upgrading to.  However, if we upgrade to rxjs 6 we also have to update to redux-observable 1.0 and that requires us to upgrade redux as well.  I can't commit to that effort, so we'll pin rxjs to 5.x in kayenta and include it in the bundle.