refactor(titus): Make titus package independent
This is the start of a new structure for all spinnaker packages such as `titus`, `core`, `amazon`.

The overall idea is to make each package independent by making them specify their own npm dependencies
and use `rollup` for building the package distribution. This helps us in migrating towards a
npm-workspace-enabled mono-repo and use tools such as lerna to manage local development and package
updates.

We're introducing a new `tsconfig.base.json` file to define the base configuration. Once all the packages
are migrated, then we can get rid of the top-level `tsconfig.json` file completely.