feat(kubernetes/job): Return the most recent pod name instead of the list of pod status (#5360)
* feat(kubernetes/job): Return the most recent pod name instead of the list of pod status

Kubernetes jobs may be executed by many pods, and each pod status is saved to the pipeline execution context. If the number of pods is large enough, the pipeline execution context may grow big enough to make orca crash.
The only place where pod status is needed is in the UI to extract the logs of the most recently created pod. A new field `mostRecentPodName` was added to the response that replaces the list of pod status. The full list of pod status may still be returned by setting the new configuration flag `kubernetes.jobs.detailed-pod-status: true` in `clouddriver.yml`.

* feat(kubernetes/job): Changed default behavior to be backwards compatible