fix(kubernetes): support options.cascading in delete manifest pipelines (#5426)
* test (kubernetes): demonstrate current behaviour of delete options in delete manifest stage

Note that the cascading checkbox in Deck isn't currently connected
because the value of the options.cascade field in the pipeline does not
have any affect on the resulting deletionOptions.orphanDependants flag.

* fix(kubernetes): support options.cascading in delete manifest pipelines

Previously, setting options.cascading in a delete manifest pipeline was
a no-op. Now, it's another way (in addition to options.orphanDependents)
to control whether kubernetes performs a cascading delete. If both
orphanDependents and cascading are specified, orphanDependents takes
precedence

Co-authored-by: Chuck Lane <clane@salesforce.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>