refactor(kubernetes): move Kubernetes Accounts config to its own class (#5444)
* refactor(kubernetes): move Kubernetes Accounts config to its own class

This is required because of side effects from the custom binding logic introduced in the PR https://github.com/spinnaker/clouddriver/pull/5125. The custom property binding fix was specifically targetted to improve account loading times from the config file. This works well if the config class, to which the config file is being deserialized, only has accounts. But if it has any other properties, these would need to be explicitly handled by the custom binding logic. This becomes an issue if there are a lot of such properties. Moreover, these properties would need to be co-located in the same file where the accounts are defined.

By refactoring Accounts into its own class, we solve the above issues since only the accounts property is now being custom bound - all other properties are loaded by spring in the default manner. This means that accounts and these properties need not be co-located anymore.

* refactor(kubernetes): remove @Primary notation

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>