fix(core/application): Fix for `requiredGroupMembership` breaking UI (#9610)
Fixes a problem that can occur for some old apps that still have `requiredGroupMembership` set on the model in front50 (instead of or in addition to `permissions`). This breaks the rendering with a `$digest already in progress` error message. Removing the `$digest()` call and instead wrapping the code in an `$evalAsync()` callback solves the issue.

Co-authored-by: Chris Thielen <christopherthielen@users.noreply.github.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>