Sql caching duplicates (#5481)
* test(cache): print test name to standard out to make test boundaries more obvious

* test(cache): demonstrate current cache behavior when adding multiple items with the same key

* fix(cache/sql): improve behavior when attempting to cache an item with a duplicate id

Previously, given items to cache in the same order, the contents of the cache would cycle
through the items.  With this change, the contents of the cache is predictable -- the
first item.

It's still not ideal for legitimately different items to have the same id (e.g. two AMIs
with the same name in the same account), but at least now the cache stays constant instead
of changing, say, each time an ImageCachingAgent runs.

* feat(cache/sql): surface attempts to cache an item with a duplicate id

* tone down message about caching duplicates to a warning

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>