fix(travis): Avoid a potential NPE when retrieving latest builds (#964)
After upgrading from Travis Enterprise 2 to 3, we ran into an issue where the build was `null` for some jobs. I believe this  is a bug in Travis, but I'm adding a check here to just avoid a NullPointerException that blocks the Travis provider from fetching builds if this issue should ever occur again.