fix(plugins): Use getExtensionClass() instead of getClass() for ExpressionFunctionProviders (#864)
ExpressionFunctionProvider extends SpinnakerExtensionPoint, enabling SpEL expressions to be contributed by plugins. However, when the functions are being registered, the class of the ExpressionFunctionProvider is used - for ExpressionFunctionProviders coming from a plugin, this is a proxy class. In order for this to work correctly, the getExtensionClass() method should be used instead to resolve the appropriate class (proxy class or not).

Co-authored-by: David Byron <82477955+dbyron-sf@users.noreply.github.com>