fix(retrofit): handle HttpStatus.NOT_FOUND with a SpinnakerHttpException instead of a NotFoundException (#911)
since SpinnakerHttpException includes details from the http response in their exception
message.  Since these exceptions eventually make it to spinnaker pipeline execution json
and the spinnaker UI, these details can be important.  That's why this commit is
considered a fix -- it makes incomplete information more complete.

It’s simple and clear for code to be able to check for a NotFoundException when it needs
to. However, the current constructor for NotFoundException doesn’t include the message
from the http response and so that information is lost. It would be possible to build a
new constructor for NotFoundException that takes a RetrofitError and re-use the
functionality in SpinnakerServerException to parse the http response, but that doesn’t
help one of the other things that code calling other microservices sometimes needs to do —
catch exceptions, add to the message, and re-throw them. For that, when NotFoundException
is a special case, it’s more work than catching only, say, SpinnakerHttpException, so we
land on this approach.

Note that this is potentially a breaking change, as code that's catching NotFoundException
needs to change to catch SpinnakerHttpException and examine the status code.