fix(core): ensure that users are not warned of their success (#9794)
The ternary as written would allow a singleton list containing a single
`null` to propagate throughout the render method of this warning
message. This would result in a menacing "warning: no reason provided"
message appearing on all stage executions.

The assignment now states that a singleton list will only be provided
iff `message` is non null and the `messages` array is empty.