feat(spin): Container improvements, non-root, exec spin, :latest tag (#331)
* feat(spin): Run container as non-root user and exec spin

Following the principle of least privilege we should not run as root
user. There is no need.
Match other Spinnaker services and run as a high user id.

Additional containers generally run the application process as
the entrypoint rather than a shell.
This makes performing commands simpler as we can tack on any arguments
to the end of the image like so:

```
$ docker run --rm -it spin:latest --version
version dev
```

The `VERSION` build arg now defaults to `dev` otherwise an empty VERSION
string overwrites default in Go and results in the `--version` flag
disappearing.

* feat(ci): GHA - tag master built container as ":latest"

We have a convention in Spinnaker land of tagging containers with
`master-latest-unvalidated`.
`master` comes from the branch (master|release-*) and for `master` could
potentially be dropped.
`unvalidated` is used because of nightly integration tests which haven't
been used with `spin` before.

All in all push `:latest` tag on successful CI run for master commits.
Standard UX in container-land.