Fix for size (#4242)
* Most K8S deployments do not use the outputs in downstream stages. So, if it is not used
users can skip the output of the Deployment Manifest and Run Job (Manifest) stage by
specifiying the attribute noOutput set to false in the stage json.

This significantly reduces the execution context being saved in redis/databse or
sent to the browser.

* feat(kubernetes): attribue to explicitly skip the outputs section in Deployment Manifest or Run Job stage

Most K8S deployments do not use the outputs in downstream stages. So, if it is not used
users can skip the output of the Deployment Manifest and Run Job (Manifest) stage by
specifiying the attribute noOutput set to false in the stage json.

This significantly reduces the execution context being saved in redis/databse or
sent to the browser.

Co-authored-by: Nirmalyasen <nirmalya@opsmx.com>
Co-authored-by: Austin Thao <austinthao@Austins-MacBook-Pro.local>
Co-authored-by: Cameron Motevasselani <cmotevasselani@gmail.com>