fix(kubernetes): Pin Debian Package version of kubectl as well (#5685)
Currently installing `spinnaker-clouddriver` results in `kubectl`
version 1.23.5 being installed.

This commit changes behaviour to install the same version as currently
specified in the `Dockerfile.slim|ubuntu` files rather than the latest stable
version.

FWIW the script doesn't upgrade `kubectl` if it is already installed which is
great as some users may manage this component separately via Ansible/etc.

However this does present a problem if/when an update to a later clouddriver
version would normally install with a newer `kubectl`.
No changes in this regard.