feat(tuning): Bump up default sizes (#968)
* feat(tuning): Bump up default size on max Connections

At any kind of scale, some of these settings are rather small.  Let's look at bumping some defaults to a TOUCH more conservative levels.  At least in ways that should NOT be 'negative" impact wise.  These are "max" vs. "min".  Open to discussion as well on the above!

* feat(http): Remove some weaker cyphers

Removes a couple of weaker cyphers that probably should NOT be supported by default. Bumps several HTTP versions.  ALSO removes TLS1.1 as a supported tls version.

* chore(conf): Adjust timing to be more clear

Makes it easier to read that it's 5 minutes when time units are minutes vs. seconds to milis.

Co-authored-by: Matt <6519811+mattgogerly@users.noreply.github.com>

Co-authored-by: Matt <6519811+mattgogerly@users.noreply.github.com>