fix(dependencies): update ant to remove CVE's (#951)
ant 1.9.15 is resolved transitively through Groovy.
Updating this removes CVE-2021-36374, CVE-2021-36373, CVE-2020-11979, and CVE-2020-15250.

At Armory, we have had the 1.10.+ config since 6/4/20, meaning we have run 1.10.8 to 1.10.12 since they came out. So this is a pretty safe change. We are just moving our own overrides to open source to fix CVE's for everyone.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>