fix(artifacts): Expected Artifacts should be trigger specific (#4322)
* fix(artifacts): Expected Artifacts should be trigger specific

When setting up multiple triggers with different expected artifacts, triggering the pipeline fails if _any_ defined expected artifact is missing, even if it is defined on another trigger. This fix filters the list of expected artifacts to make sure only the artifacts that are relevant for the current trigger is evaluated.

* Fix tests in ArtifactUtilsSpec

Signed-off-by: Jørgen Jervidalo <jorgen.jervidalo@schibsted.com>

* Add expectedArtifactIds to pipeline triggers, add more tests and fix some tests that failed

Signed-off-by: Jørgen Jervidalo <jorgen.jervidalo@schibsted.com>

* Fix typo

Signed-off-by: Jørgen Jervidalo <jorgen.jervidalo@schibsted.com>

Signed-off-by: Jørgen Jervidalo <jorgen.jervidalo@schibsted.com>