feat(secrets): Add credentials masking in JSON (#992)
This adds a new `Sensitive` annotation to explicitly mark fields as
sensitive. A sensitive field will not be serialized to JSON unless its
value is a `secret://...` URI to avoid leaking credentials. This
serializer also provides a configurable regular expression to match
against `CredentialsDefinition` property names to mark them as likely
sensitive which applies masking in addition to logging a warning so that
the affected class can be updated.