feat(pipeline/kubernetes): Introduce validator for kubernetes blue/green traffic management strategy (#1176)
Introduce blue/green strategy as this is a terminology used more often in the industry compared to the Neflix specific phrasing. First step is to deprecate redblack. When creating new pipelines using fron50 api, if traffic management is enabled and redblack is used a warning message is logged.

In the future, after the redblack will be completely remove a validation error will be returned