feat(pubsub): add support for links in pubsub triggers (#1235)
It is nice how how Jenkins triggers display a link back to the build that triggered the pipeline. This keys off the link and linkText properties in the Trigger object. I would like to see this same behavior supported for pubsub triggers. This change allows senders of pubsub events to add link and linkText in the trigger message payload, and thus set a link that will be displayed in the pipeline execution UI.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>