refactor(sql): Optimize read queries for getUserPermission (#1003)
This fixes a couple issues in the existing implementation of getUserPermission. First, the account manager flag is queried at the same time as the admin flag (which is now fixed thanks to better understanding of the jOOQ API). Secondly, the functions to fetch resources for a given user were doing a natural join in Kotlin rather than offloading the request to the database to join.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>