fix(travis): Enable legacy log fetching (#1075)
* fix(travis): Enable legacy log fetching

We've had lots of trouble with the log_complete feature Travis added a while back on our request. It was supposed to solve our issues with figuring out when logs are actually available (which can be up to several minutes after the build is complete) - but basically it can't be trusted. We've seen cases where it turns to true even before the build was completed (how can the logs be done then? 😆), and also in the opposite case, it can return false for a long time even though the logs are in fact complete. So it's better to just fall back to the old way of checking log completion (with some improvements). We've been running this internally in Schibsted for several months, and have not seen any issues with log fetching during that time - a huge improvement from earlier!

* Add feature flag for log_complete (defaulting to off)