feat(web): Add X-SPINNAKER-* optional data to HTTP response header (#1610)
* test(web): Add unit test for RequestIdInterceptor

* feat(web): Refactor RequestIdInterceptor to be able to set additional X-SPINNAKER header info besides request ID to response header, and renaming it to ResponseHeaderInterceptor to better reflect its functionality. The motivation for this enhancement is to allow better correlation of API interaction with users making these requests.

A new configuration property "interceptors.responseHeader.fields" is added for configuring the X-SPINNAKER-* header fields to be added to the response header by the interceptor. This property takes a comma delimited string or yaml list of the header fields, for example to add X-SPINNAKER-REQUEST-ID and X-SPINNAKER-USER to the response header:

interceptors.responseHeader.fields=X-SPINNAKER-REQUEST-ID,X-SPINNAKER-USER

or in yaml form:

interceptors:
  responseHeader:
    fields:
      - X-SPINNAKER-REQUEST-ID
      - X-SPINNAKER-USER

When this configuration property is not defined, the interceptor defaults to adding X-SPINNAKER-REQUEST-ID field to the response, which is the current behavior.