fix(plugins): Enable RemotePluginInfoReleaseCache bean only when the property matches the condition (#1041)
Last week during Github's downtime, because the RemotePluginInfoReleaseCache runs in the background every 1 minutes, it forced the plugins to be reloaded(downloaded again from Github), action which failed because the Github keys were rotated due to their security leak

The refresh of the plugins is caused by the updateManager.refresh() call in the refresh method of the RemotePluginInfoReleaseCache.

As far as we can tell, the concept of remote plugins is something that is not used at all. So we can opt-in for this cache refresh bean by manually setting spinnaker.extensibility.remote-plugins.cache.enabled to true

Incident report page: https://www.githubstatus.com/incidents/x7njwb481j9b