feat(core + sql): Optimize cache refresh (#1249)
* refactor(core): use lombok in StorageServiceConfigurationProperties to reduce boilerplate

* refactor(core): set defaults for StorageServiceConfigurationProperties via setters instead of a constructor

to make it easier to add more properties

* refactor(core): change StorageServiceSupport constructor to take a whole StorageServiceConfigurationProperties.PerObjectType object

instead of individual properties to make it easier to add more properties

* refactor(core): move config validation logic closer to where it's used in StorageServiceSupport

* feat(core): add service-storage.*.cacheHealthCheckTimeoutSeconds configuration property

that defaults to previously hard-coded value of 90 seconds.

where * is all the object types in StorageServiceConfigurationProperties:

* application
* applicationPermission
* serviceAccount
* project
* notification
* pipelineStrategy
* pipeline
* pipelineTemplate
* snapshot
* deliveryConfig
* pluginInfo
* entityTags

* fix(web): clarify log message in PipelineController.update

when the id of the provided pipeline doesn't match the id of the existing pipeline

* refactor(core): move refresh logic into a separate method in StorageServiceSupport

to prepare to add synchronization

* feat(core): add service-storage.*.synchronizeCacheRefresh configuration property

that defaults to false.  When true, if multiple threads attempt to refresh the cache in
StorageServiceSupport simultaneously, only one actually does the refresh.  The others wait
until it's complete.  This reduces load on the data store.

* is all the object types in StorageServiceConfigurationProperties:

* application
* applicationPermission
* serviceAccount
* project
* notification
* pipelineStrategy
* pipeline
* pipelineTemplate
* snapshot
* deliveryConfig
* pluginInfo
* entityTags

* refactor(sql): remove unnecessary class specifiers, other small cleanups

* feat(sql): add service-storage.*.optimizeCacheRefreshes config flag

that defaults to false.  When true, for objects that support versioning, cache refreshes only query the data store for objects modified (or deleted) since the last refresh.

This uses a new function StorageService.loadObjectsNewerThan that's currently only implemented for sql data stores.

---------

Co-authored-by: Jason <mcintoshj@gmail.com>