feat(front50): call front50's GET /pipelines/triggeredBy/{pipelineId}/{status} endpoint (#4448)
* refactor(front50): introduce Front50ConfigurationProperties class

in preference to @Value annotations

* feat(front50): call front50's GET /pipelines/triggeredBy/{pipelineId}/{status} endpoint

in DependentPipelineExecutionListener when the front50.useTriggeredByEndpoint configuration property is true.  The default value is false.

The advantage of the new endpoint is that front50 only returns a (potentially very small) subset of pipelines over the wire to orca, in contrast to the current behavior where front50 returns all pipelines, and leaves it to orca to filter them.