feat(auth): Disable Fiat to allow anonymous access
This enables the installed Spinnaker to create pipelines and deploy
applications into the cluster it is installed in which is useful for
someone testing Spinnaker out for the first time or for some development
flows.

It is also a huge security risk if Spinnaker is deployed in a cluster
accessible to malicious actors.

It would be good to enable Basic Authentication for Gate and then
re-enable Fiat. TODO.