Cryptography module and OkHttp3 config updates (#1050)
* feat(crypto): Add X509IdentitySource API

This adds a new Kork module for common cryptographic operations. Primarily, this introduces the `X509IdentitySource` API for loading private keys and certificates as cryptographic identities and subsequently using them in `SSLContext` instances for configuring mutual TLS authentication. Included are two implementations for `X509Identity`: a static one containing a loaded private key and certificate chain, and a dynamic one which can be refreshed periodically to reload a key and certificate whenever the backing file changes.

Some other goodies that come with this new module:

- `SecureRandomBuilder` for creating a custom `SecureRandom` instance using a DRBG strategy (new in Java 9). This makes it simpler to configure domain-specific random number generators in a safe way.
- `PasswordProvider` interface for abstracting how the password for a keystore or private key is loaded (e.g., from a file).
- Updated version of Bouncycastle using the Java 1.8+ module.

* feat(web): Add support for refreshable mTLS identity

This integrates the new kork-crypto module into kork-web to allow OkHttp v3 clients to use a refreshable identity. To avoid duplication of config logic and allowing this feature to be enabled via a feature flag, a new `OkHttpClientCustomizer` interface and prototype bean for `OkHttpClient.Builder` are introduced analogous to Spring's `WebClientCustomizer` and `WebClient.Builder` interfaces.

As a bonus, the `OkHttpClientComponents` config class is migrated from Groovy to Java (along with moving a little bit of Groovy annotations for better IDE detection).

* refactor(web): Simplify constructors in OkHttp3ClientConfiguration

* refactor(crypto): Avoid use of deprecated SocketUtils method

* chore(crypto): Use Bouncycastle 1.70

* refactor(crypto): Move crypto test code to test fixture

This makes it simpler to reuse test code via the test fixture Gradle feature.

* test(web): Add test for refreshable https client