feat(sql): add configuration property sql.healthIntervalMillis (#1275)
that controls how often the health endpoint refreshes information when using sql.  It
defaults to 30 seconds to match the behavior before this change.

Information about cache status for each object type (e.g. pipelines, applications, etc.)
contributes to the response from front50's /health endpoint.  This change makes it
possible to reduce the likelihood that the cache has refreshed, but the response from the
/health endpoint doesn't reflect that.