fix(core): tolerate items with null ids (#1276)
* test(core): demonstrate current behavior with a pipeline with a null id in the database

@W-12733072

* fix(core): filter out items with null ids in StorageServiceSupport.fetchAllItems

Even if it's not possible to get pipelines without ids into the database via an API
endpoint, there's nothing stopping a wayward sql query or a bug elsewhere from getting
them there.  So, it's helpful for the code to handle this.

* fix(core): tolerate items with null ids in StorageServiceSupport.findById

Even if it's not possible to get pipelines without ids into the database via an API
endpoint, there's nothing stopping a wayward sql query or a bug elsewhere from getting
them there.  So, it's helpful for the code to handle this.

@W-12733072

* fix(core): filter out items with null ids in StorageServiceSupport.fetchAllItemsOptimized

Even if it's not possible to get pipelines without ids into the database via an API
endpoint, there's nothing stopping a wayward sql query or a bug elsewhere from getting
them there.  So, it's helpful for the code to handle this.

@W-12733072

* refactor(core/test): combine null id tests in PipelineControllerTck

since the expectations are so similar whether optimizeCacheRefreshes is true or false

@W-12733072

* fix(core): throw an IllegalArgumentException from StorageServiceSupport.update

to prevent items with null ids from getting into the database

@W-12733072

* test(core): demonstrate behavior of StorageServiceSupport.bulkImport of a pipeline with a null id