perf(orca/executions): add pipelineNameFilter to /applications/{application}/pipelines endpoint
and /v2/applications/{application}/pipelines endpoint as well.

This adds the pipelineNameFilter to the endpoints which get pipelines by application
name. Right now this filter is present in gate, but gate does the filtering with the
parameter, instead of passing it through to orca. This is inefficient as there is no need
for orca to send over all of the pipeline execution information for an application, and
then have gate do the filtering. With this functionality here, we can change gate to
simply pass the parameter through.