perf(front50/pipelines): pipelineLimit query param on /pipelineConfigs
Adds a pipelineLimit query param on the /{application}/pipelineConfigs endpoint. This parameter ensures that the number of pipelines returned from the endpoint never exceeds the number specified by pipelineLimit.

Combined with similar changes to gate and orca, this paves the way for a UI change to limit the number of pipeline configs and executions on the executions page.