fix(front50/pipelines): fix bug rewriting pipeline index in cache
Fixes a bug that rewrites the pipeline index value of the pipelines in the front50 cache. The previous code that set p.setIndex(i) was actually rewriting the index values of the pipelines in the cache, because the p is actually a reference to the underlying memory of the cache. To fix, we do a shallow clone of the pipeline objects that we are going to return, and set the index on the cloned objects.

This bug became apparent with the pipelineLimit query parameter, as subsequent calls to the same endpoint with the same limit value were returning different results. This also adds a specific test case for that very scenario, to ensure that front50 always uses the index values that exist in the db.