perf(execution): Add a retrieve method with requireLatestVersion
Adds a retrieve method with the requireLatestVersion option. This option
is useful in use cases where the execution repository has a read
replica and retrieve()'s caller does not need the latest version of the
PipelineExecution. In these cases, the caller can retrieve the
PipelineExecution currently stored in the read replica. This is a
much cheaper operation than querying the main writer instance.