SpinnakerHttpException more properties (#1093)
* refactor(retrofit): explicitly verify that the RetrofitError has a non-null response in the SpinnakerHttpException constructor

to remove the need for null checks elsewhere

* refactor(retrofit/test): store url in a variable in SpinnakerHttpExceptionTest

to pave the way for verifying a new url accessor on SpinnakerHttpException

* refactor(retrofit/test): Retrofit.Builder adds a trailing slash to the baseUrl if not present

Add one ourselves so there aren't as many moving parts

* feat(retrofit): add SpinnakerHttpException.getUrl()

so it's available to exception handling code (e.g. the Spinnaker*Exception equivalent of orca's RetrofitExceptionHandler)

* refactor(retrofit): add a responseCode member to SpinnakerHttpException

to reduce use of the response and retrofit2Response members outside of constructors.

Simplify getMessage() now that the retrofit1 and retrofit2 cases are the same.

* feat(retrofit): add getReason() accessor to SpinnakerHttpException

so it's available to exception handling code (e.g. the Spinnaker*Exception equivalent of orca's RetrofitExceptionHandler)