chore(build): empty commit to prepare for release-1.32.x
This puts the HEAD of master at a different commit than the first tag for the
release-1.32.x branch, so the github action knows to create autobump PRs targetting the
release-1.32.x branch.  Without this, v1.42.0 is at the HEAD of both master and
release-1.32.x and the release github action doesn't know what to target.