feat(orca/execution): Add RedisExecutionUpdateTimeRepository
Creates a redis-backed ExecutionUpdateTimeRepository which persists
the latest update time for a particular pipeline or stage execution.
This repository will help ensure that the execution that we fetch
from an ExecutionRepository is up-to-date in situations where
consistency is not guaranteed. E.g. when using a read replica,
there is often replication lag from the primary instance to the read
replica, which can cause the execution stored in the read replica to
be out of date.