feat(retrofit): add SpinnakerServerExceptionHandler (#4524)
* test(queue): demonstrate behavior of RunTaskHandler when a task throws a RetrofitError

with a fix of a corner case in RetrofitExceptionHandler along the way.

This paves the way to demonstrate similar behavior when a task throws a SpinnakerServerException

* refactor(retrofit): split retry logic from RetrofitExceptionHandler into BaseRetrofitExceptionHandler

to pave the way to share it with a new SpinnakerServerExceptionHandler class

* refactor(retrofit): change BaseRetrofitExceptionHandler.shouldRetry to take an Exception instead of only RetrofitError

The logic in the implementation is still specific to retrofit as it looks for
retrofit-specific annotations in the stack trace.  This way though,
SpinnakerNetworkExceptions caused by e.g. IOExceptions (e.g. with no underlying
RetrofitError in the causal chain) are considered retryable when they happen in an
idempotent retrofit request.

* refactor(retrofit): change BaseRetrofitExceptionHandler.shouldRetry's kind argument to a String

which decouples children of BaseRetrofitExceptionHandler from retrofit1.  This makes it easier to
remove retrofit1 from the classpath once we move to retrofit2.

* feat(retrofit): add SpinnakerServerExceptionHandler

to provide similar functionality when tasks throw SpinnakerServerException that
RetrofitExceptionHandler provides for RetrofitException.  Specifically, this means:

* retry behavior is the same with SpinnakerServerExceptionHandler as it is with RetrofitExceptionHandler
* the information in the pipeline execution context is similar
* the way errors appear in the UI with SpinnakerSeverExceptionHandler is a combination of
  DefaultExceptionHandler, which is an improvement over RetrofitExceptionHandler
  (e.g. http response codes and URLs are present)

* feat(retrofit): add SpinnakerServerExceptionHandler bean

so there's a SpinnakerServerExceptionHandler available to RunTaskHandler