fix(artifact): Handle exceptions better in artifact store (#1103)
* fix(artifact): Handle exceptions better in artifact store

Prior to this commit, when the head object in s3 artifact store throws
an exception in (de)serializer, we get a pretty bland exception back

```json
{"timestamp":"2023-09-25T17:54:02.550+00:00","status":500,"error":"Internal
Server Error"}
```

With this commit we will now see:
```json
{"timestamp":"2023-10-02T23:21:42.183+00:00","status":500,"error":"Internal Server Error","exception":"com.netflix.spinnaker.kork.artifacts.artifactstore.exceptions.ArtifactStoreIOException","message":"com.netflix.spinnaker.kork.exceptions.SpinnakerException: Failed to query artifact due to IAM permissions either on the bucket or object"}
```

The fix here was to intercept any `Exception` from the artifact store
calls and ensure we rethrow a `ArtifactStoreIOException`. The reason why
we rethrow an `IOException` is due to what jackson does with exceptions.
If we do not throw an `IOException`, then jackson will wrap it in a JSON
exception which is confusing to end user.

Signed-off-by: benjamin-j-powell <benjamin_j_powell@apple.com>

* Update kork-artifacts/src/test/java/com/netflix/spinnaker/kork/artifacts/artifactstore/EmbeddedArtifactSerializerTest.java

---------

Signed-off-by: benjamin-j-powell <benjamin_j_powell@apple.com>
Co-authored-by: Benevolent Benjamin Powell <benjamin_j_powell@apple.com>