feat(orca/execution): Return status codes from ExecutionMapper.map
Refactors ExecutionMapper.map to return a data class with two fields:
1. The existing Collection<PipelineExecution>
2. A new ExecutionMapperResultCode that provides details about the result
of the map() operation.
ExecutionMapperResultCode is needed now that the SqlExecutionRepository
can use read replicas, which introduces more complex failure cases than before.
Previously, ExecutionMapper.map() would return just a Collection. If the collection
is empty, that means that the provided ResultSet did not contain any rows. With
the implementation of read pools, map() can be given a valid ResultSet but still
return an empty Collection. For example, if an execution in the ResultSet is not
a new enough version, map() will return an empty Collection along with
ExecutionMapperResultCode.INVALID_VERSION.

The new result codes give callers the info to perform follow-up operations as needed.