feat(retrofit): move url attribute from SpinnakerHttpException to SpinnakerServerException (#1110)
* refactor(retrofit/test): cleanup tests in kork-retrofit

to use fluent assertions and other small tweaks

* test(retrofit): add assertions on SpinnakerHttpException's url

to prepare to move url to SpinnakerServerException

* refactor(retrofit/test): remove duplicate test from SpinnakerRetrofit2ErrorHandleTest

testRetrofitSimpleSpinnakerServerException actually generates a SpinnakerNetworkException and so it's effectively the same test as testRetrofitSimpleSpinnakerNetworkException.

* test(retrofit): test behavior when a retrofit endpoint generates a SpinnakerNetworkException

* feat(retrofit): move url attribute from SpinnakerHttpException to SpinnakerServerException

so it's available to exception handling code that currently has access to the url via RetrofitError.

* refactor(retrofit): change Spinnaker*Exception constructors to take a Request

instead of a url, to pave the way for storing other attributes of requests (e.g. the http request method)

* fix(retrofit): add javadoc, and properly set retryable to false in the new SpinnakerConversionException constructors

* refactor(retrofit): use @Getter for url in SpinnakerServerException