chore(dependencies): Autobump korkVersion (#1188)
* chore(dependencies): Autobump korkVersion

* fix(web): fix exception handling in JenkinsBuildMonitor.processBuildsOfProject

The jenkinsClient inside JenkinsService comes from JenkinsConfig, wihch, as of
https://github.com/spinnaker/igor/pull/1130, uses SpinnakerRetrofitErrorHandler.
Therefore, the getBuilds calls inside the try/catch block inside processBuildsOfProject
throws a SpinnakerServerException.

https://github.com/spinnaker/igor/pull/1130 updated the 'should continue processing other
builds from a master even if one or more build fetches fail' test to throw a
SpinnakerServerException with a RetrofitError as its cause instead of a RuntimeException
with a RetrofitError as its cause.  With https://github.com/spinnaker/kork/pull/1110, the
SpinnakerServerException no longer has a RetrofitError as its cause.  It has the "mock
root cause" exception as its cause.  But, because
https://github.com/spinnaker/kork/pull/1110 adds a url attribute to
SpinnakerServerException, the cause is no longer necessary for including a url in the
"Error communicating with jenkins" log message, and this commit updates the logic
accordingly.

* refactor(web/test): clean up exception mocking in JenkinsBuildMonitorSpec

remove unused runtimeException and rename retrofitEx since it's no longer a retrofitEx

---------

Co-authored-by: root <root@2f91e8016c3c>
Co-authored-by: David Byron <dbyron@salesforce.com>