fix(artifacts): Parent and child pipeline artifact resolution (#4575)
This fix allows us to revert https://github.com/spinnaker/orca/pull/4397
since that makes retrieving bound artifacts actually more strict for
most use cases. So this commit addresses the root cause that that PR was
trying to implement.

When a parent pipeline triggers a child pipeline, the child pipeline
should be able to resolve incoming artifacts properly. Right now that
does not happen because `expectedArtifactIds` is needed in the trigger
section, otherwise no resolution will occur.

This change fixes the parent payload to include `expectedArtifactIds` by
viewing the child pipeline's top level expected artifacts and finding
the intersection between expected artifacts and artifacts to be sent.
With that intersection we can grab the expected artifact ids and
properly set the the necessary field to allow for child pipelines to
ingest the artifact.

Signed-off-by: benjamin-j-powell <bjp@apple.com>
Co-authored-by: Benevolent Benjamin Powell <benjamin_j_powell@apple.com>