config(crypto): Update the default TLS cipher suites (#1112)
This adds a common `CipherSuites` class for listing various profiles of cipher suites for use with TLS. The default cipher suites and TLS versions are updated to a slightly more secure version of the intermediate compatibility baseline from the Mozilla recommendations. These settings are used for both Tomcat and OkHttpClient. Of particular note, this adds in the required ciphers in TLS 1.3 to improve compatibility with modern clients and servers.