perf(sql): Simplify sql query in retrievePipelineExecutionDetailsForApplication
Since the query already specifies ids, the additional where condition for application is
redundant and can be removed.

Before the change, the query uses pipeline_application_idx.  After it uses the primary key (id).

While we're at it, switch to use the read pool since replication lag is acceptable here.
It's only used by gate / the UI to retrieve pipelines to display.