feat(kubernetes): Support multiple kubectl versions (#6134)
* feat(kubernetes): Support multiple kubectl versions

Currently it is hard to support clusters that runs different Kubernetes versions. Because kubectl executables can be defined per cluster, we can overcome this limitation by installing multiple kubectl versions, and then define which binary to use in the cluster config.
The kubectl versions are stored without patch version, so that we can upgrade to newer versions without having to change the configuration. We also symlink `kubectl` to the current default version to not break existing configuration.

To opt in to a newer version, define it like this in the cluster config: `kubectlExecutable: /usr/local/bin/kubectl-1.28`
E.g.:
```
kubernetes:
  enabled: true
  primaryAccount: my-cluster
  accounts:
    - name: my-cluster
      kubeconfigFile: configserver:k8s/my-cluster.yml
      cacheThreads: 1
      cachingPolicies: []
      checkPermissionsOnStartup: true
      customResources: []
      dockerRegistries: []
      kinds: []
      liveManifestCalls: false
      namespaces: []
      oAuthScopes: []
      omitKinds: []
      omitNamespaces: []
      onlySpinnakerManaged: false
      providerVersion: V2
      kubectlExecutable: /usr/local/bin/kubectl-1.28
```

* Be less verbose

* Upgrade from KIND v0.11.1 to v0.20.0

This includes adding Kubernetes 1.29 and removing 1.19 and 1.20

* Select correct kubectl version based on Kubernetes version

To adhere to the Kubernetes version skew policy

* Fix test that failed with kubect v1.26 and above

* Fix kubectl download path

* We can't depend on the jsonpath implementation in kubectl :(

* Fix import order

* Fix comment

* Replace switch with map

* Fix symlink

* Move all of the kubectl version selection code out to the github actions definition

All kind versions will be tested with the default kubectl version (currently 1.22.17), as well as the version that corresponds to the k8s version.

* Second try on the matrix definition

* Make the titles in GHA a bit more readable

* Move the constant back to the original position

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>