fix(artifacts): Addresses a couple things regarding artifact store (#1145)
* fix(artifacts): Updates Artifact Store README

This commit updates the README refect the changes in
https://github.com/spinnaker/kork/pull/1120

Also adds a Rosco/Helm section to describe the new `expandOverrides`
field

Signed-off-by: benjamin-j-powell <bjp@apple.com>

* feat(artifacts): Move scheme to ArtifactReferenceURI

This moves scheme from the artifact URI builders to the URI object
instead. Also moves the `isArtifactReference` from artifact stores to
the URI class instead.

Signed-off-by: benjamin-j-powell <bjp@apple.com>

* feat(artifacts): Add HelmConfig and ensure properties aren't null

This commit adds the new HelmConfig which allows for users to turn on
expanding overrides. This will inspect overrides values, and if it is an
reference URI, it will expand and use that instead of the raw URI.

This also updates the properties, `helmConfig` and `s3` to no longer be
`null`.

Signed-off-by: benjamin-j-powell <bjp@apple.com>

---------

Signed-off-by: benjamin-j-powell <bjp@apple.com>
Co-authored-by: benjamin-j-powell <bjp@apple.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>