feat(artifact): Introduces the expansion of artifact store references in parameters  (#1063)
* feat(artifact): logic to handle reference URIs in overrides

This commit updates the Helm util helpers that we have to include the
artifact store. This will allow us to inspect the raw strings and check
to see if we need to expand them.

Further, I opted to not utilize the deserializer this time, since this
problem seemed very specific to Helm. If that changes, we can always
just use the deserializer

Signed-off-by: benjamin-j-powell <bjp@apple.com>

* feat(artifacts): Updates to tests due to Helm util constructor changes

Signed-off-by: benjamin-j-powell <bjp@apple.com>

* fixup! feat(artifact): logic to handle reference URIs in overrides

Signed-off-by: benjamin-j-powell <bjp@apple.com>

---------

Signed-off-by: benjamin-j-powell <bjp@apple.com>
Co-authored-by: benjamin-j-powell <bjp@apple.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>