chore(dependencies): Autobump orcaVersion (#1018)
* chore(dependencies): Autobump orcaVersion

* fix(test): Update swagger-ui endpoint with springfox 3.0.0 upgrade

With upgrade of springfox from 2.9.2 to 3.0.0, a breaking [change](http://springfox.github.io/springfox/docs/current/#changes-in-swagger-ui) has been introduced in the form of new endpoint for swagger-ui (/swagger-ui/index.html). This change caused the `SwaggerTest.swaggerUiIsPresent` test to fail. In order to fix this test, updating the swagger-ui endpoint as `/swagger-ui/index.html`.

---------

Co-authored-by: root <root@9d081cf5382b>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: j-sandy <30489233+j-sandy@users.noreply.github.com>