feat(orca/sql): Use ExecutionUpdateTimeRepository in SqlExecutionRepository
When the config flag execution-repository.sql.read-replica.enabled is set,
use the RedisExecutionUpdateTimeRepository to track the latest known
update time for a particular execution. When retrieving an execution,
first attempt to use the read replica. If the read replica does not
contain an execution newer than the latest update time, then reject
the result and query from the primary instance instead. An execution
is considered to be newer than the latest update time if it meets
the following conditions:
    1. The update time of the execution in the database, denoted by the
       value of its updated_at column, is newer than or equal
       to the required update time
    2. If compression is enabled, the update time of the compressed
       execution, denoted by the updated_at column, is newer than or equal
       to the required update time
    3. If the execution contains stages, the update time of each stage,
       denoted by the updated_at column, is newer than or equal to the
       required update time