refactor(igor): use SpinnakerRetrofitErrorHandler with IgorService (#4666)
* refactor(igor): use SpinnakerRetrofitErrorHandler with IgorService

This PR lays the foundational work for upgrading the retrofit version to 2.x, specifically focusing on refactoring the exception handling for IgorService.

There are no behaviour changes detected with these code changes and hence no additional tests added to demonstrate the functionalities. Few existing tests are modified to align it with the new changes.

* refactor(keel): Cleanup RetrofitError in ImportDeliveryConfigTask

There are 2 APIs invoked inside the try block:
1. Keel Service API - keelService.publishDeliveryConfig() : With the addition of the commit - c41792295d4b2bba289ab131699cc7b061ec627e, the keelService uses SpinnakerRetrofitErrorHandler.
2. Igor Service API - scmService.getDeliveryConfigManifest() : With the addition of the commit - d430b75efeced7eb1f8d64eff6f0a160210eb2e6, the igor service uses SpinnakerRetrofitErrorHandler.

and hence with the above mentioned commits the catch block with RetrofitError and the handler logics associated with it becomes irrelevant.