fix(orca/sql): Enforce that retrieved pipelines have
the correct number of stages. Orca can inject additional stages
(called synthetic stages) during runtime. In a system with replication
lag, it is possible for a synthetic stage to be missed when retrieving
the list of stages that belonging to a pipeline. To prevent this from
happening, use the ReplicationLagAwareRepository to enforce the
number of stages that belong to a pipeline execution. In addition to
checking that every retrieved pipeline and stage are up-to-date, also
verify that every pipeline has the correct number of stages.