refactor(front50): use SpinnakerRetrofitErrorHandler with Front50Service (#4623)
* refactor(front50): use SpinnakerRetrofitErrorHandler with Front50Service

This PR lays the foundational work for upgrading the retrofit version to 2.x, specifically focusing on refactoring the exception handling for Front50Service.

There are no behaviour changes detected with these code changes and hence no additional tests added to demonstrate the functionalities.
Few existing tests are modified to align it with the new changes.

* refactor(applications): Cleanup RetrofitError in DeleteApplicationTask

There are 3 APIs invoked inside the try block:
1. Keel Service API - keelService.deleteDeliveryConfig(): With the addition of the commit - c41792295d4b2bba289ab131699cc7b061ec627e, the keelService uses SpinnakerRetrofitErrorHandler.
2. Front50 Service APIs - front50Service.get() and front50Service.delete(): With the addition of the commit - a44b5bbc84938eb58731ea9c813d702750026fb6, the front50 service uses SpinnakerRetrofitErrorHandler.

and hence with the above mentioned commits, the catch block with RetrofitError becomes irrelevant.

* refactor(igor): Cleanup RetrofitError in GetCommitsTask

There are total of 5 API calls happening inside the try block:
1. Igor Service API - scmService.compareCommits(): With the addition of the commit - d430b75efeced7eb1f8d64eff6f0a160210eb2e6, the igorService uses SpinnakerRetrofitErrorHandler.
2. Clouddriver APIs (OortService) - cloudDriverService.getByAmiId() - This is invoked twice in the block, and oortService.getServerGroupFromCluster(): With the addition of the commit - 84a7106c512ceada65d7799f527a956262d964d5, the oortService uses SpinnakerRetrofitErrorHandler.
3. Front50 API - front50Service.get(): With the addition of the commit - a44b5bbc84938eb58731ea9c813d702750026fb6, the front50Service uses SpinnakerRetrofitErrorHandler.

and hence with the above mentioned commits, the catch block with RetrofitError becomes irrelevant.

* refactor(test/clouddriver): Cleanup impractical tests in TrafficGuardSpec

The API call 'front50Service.get(application)' mocked/stubbed to return null value is not realistic, because if the application is not available , then this API would throw a 404 HTTP error and thats already handled in the same class: https://github.com/spinnaker/orca/blob/7e923cc05af0f66b952245722aa69c667f685c5c/orca-clouddriver/src/test/groovy/com/netflix/spinnaker/orca/clouddriver/utils/TrafficGuardSpec.groovy#L464.

and hence this test is cleaned up.

* refactor(test/applications): Cleanup impractical tests in UpsertApplicationTaskSpec

The API call 'front50Service.get(application)' mocked/stubbed to return null value is not realistic, because if the application is not present, then this API would throw a 404 HTTP error.

and hence this test is cleaned up.

* refactor(test): adjust UpsertApplicationTaskSpec tests to use realistic behavior of Front50Service.get

when an application isn't found.  It doesn't return null in this case, it throws an exception.

Restore previously removed "should create an application in global registries" test since
it was providing useful coverage.

---------

Co-authored-by: Pranav-b-7 <pranav.bhaskaran@opsmx.com>
Co-authored-by: David Byron <dbyron@salesforce.com>