fix(clouddriver): Handle the Retrofit conversionError in ClusterSizePreconditionTask (#4708)
* fix(clouddriver): Handle the Retrofit conversionError in ClusterSizePreconditionTask

The oortService already uses SpinnakerRetrofitErrorHandler as part of the commit: 84a7106c512ceada65d7799f527a956262d964d5, so constructing the conversionError with RetrofitError would result in unexpected behaviour.

Wrapping up the conversion error into SpinnakerConversionError will fix this bug.

* test(clouddriver): Test to verify conversion exception in ClusterSizePreconditionTask

The expectation is that the oortService APIs must throw only Spinnaker*Exceptions as it is already using SpinnakerRetrofitErrorHandler.

This test is to ensure that any future modifications on this fix must not result in unexpected behaviour.

---------

Co-authored-by: Pranav-b-7 <pranav.bhaskaran@opsmx.com>