refactor(api): change FiatPermissionEvaluator to implement UserPermissionEvaluator (#1155)
instead of PermissionEvaluator, and mark

public boolean hasPermission(
      String username, Serializable resourceName, String resourceType, Object authorization)

as @Override.

This makes this method available to e.g. S3ArtifactStoreGetter so it can authenticate by
user.  In some pipeline execution scenarios in orca (e.g. using #fetchReference in an
Evaluate Variables stage), this is necessary since SecurityContextHolder.getContext() is
null.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>