feat(migrations): Support for migrations defined in plugins (#1458)
* feat(migrations): Support for migrations defined in plugins

The current implementation of the MigrationRunner is initialized before any migrations in plugins are defined. Even if they are defined using `@ExposeToApp`, the beans are never picked up by the MigrationRunner bean.
This fix looks up Migration beans directly from the applicationContext on every run, and will also delay the first run with 10 seconds in order to give some time to plugins to be initialized.

* Fix import order

* Add tests for MigrationRunner

---------

Co-authored-by: Jason <mcintoshj@gmail.com>