feat(batchUpdate): Pipeline config batch update (#4773)
* feat(pipeline executions/orca) : Added code to save multiple pipelines at once to sql database.

This is part of: spinnaker/spinnaker#6147.

Enhanced SavePipelineTask.java to

Added code to ensure that the SavePipelineTask.java also accepts list of pipelines json.
This method will validate all the pipelines.
This method will call the front50 service to save the pipelines list.

Enhanced Front50Service.groovy to

Added new rest api which accepts list of pipelines json.

* feat(pipelines executions/orca): Support for bulk saving pipelines

* feat(front50): Add okhttp timeouts for front50 service

Adds configuration properties for readTimeout, writeTimeout, and connectTimeout on the Front50 OkHttp service.

* bug(front50): add GroovyJson dependency

Test was failing without it. The test code that uses it was added back when we were in Groovy 2, after the upgrade to Groovy 3 I believe it's required as an explicit dependency.

* refactor(front50): remove old timeout config value from orca.yml

* refactor(front50): change RetrofitError to SpinnakerHttpException

---------

Co-authored-by: sanopsmx <sanjeev@opsmx.io>
Co-authored-by: Arifullah Pattan <apattan@salesforce.com>
Co-authored-by: David Byron <dbyron@salesforce.com>
Co-authored-by: Richard Timpson <richard.timpson@salesforce.com>