refactor(orca/sql): change ExecutionMapperResult to a sealed interface
with specific implementations for each result.  With this, we no longer need
ReplicationLagAwareResultCode, so remove it.

This helps clean up the code since we no longer need to worry about a pair of separate
values -- executions + result code.  Executions are only available on success.  This also
means that non-success result codes are easier to use.  There's no more need to build an
empty list that's not used.