feat(gate/web): add GET /pipelines/{id}/status endpoint
to return the status of a pipeline execution as a plain string (e.g. SUCCEEDED). There's a
query parameter named readReplicaRequirement that defaults to UP_TO_DATE. Valid values are
(case-insensitive):

NONE: query the read replica directly
PRESENT: wait for the execution to be present in the read replica
UP_TO_DATE: wait for the execution to be up to date in the read replica