fix(web): invoke pipeline config exception handling (#1831)
* feat(web): introduce new configuration property services.front50.applicationRefreshInitialDelayMs

which provides an initial delay in milliseconds for the thread that refreshes the
applications cache in ApplicationService

It's primarily to facilitate testing, but it seems reasonable someone might want use it
production to keep things quiet at startup.

* test(web): demonstrate current behavior of PipelineController.invokePipelineConfig

as determined by wiremock responses from front50 and orca (InvokePipelineConfigTest)

and when PipelineService.trigger throws an exception (PipelineControllerTest)

* fix(web): let exceptions during PipelineController.invokePipelineConfig bubble up

so gate's http response code more closely reflects what happened

* refactor(web): change PipelineController to use constructor autowiring
to prepare for changes to the constructor logic

* fix(web): include information from downstream services in error responses from PipelineController.invokePipelineConfig

by handling RetrofitErrors with SpinnakerRetrofitErrorHandler

As part of this PipelineController.invokePipelineConfig no longer logs its own message for
RetrofitErrors.  There's some loss of information with this, as the initiator of the
downstream communication is from no longer clear.  A subsequent commit restores this.

* fix(web): chain Spinnaker*Exceptions in PipelineController.invokePipelineConfig

so it's clear which operation is failing.  This improves both logging and gate's http response.

As part of this, remove the no-op catch and throw for NotFoundException.  With no other
more general catch block, this code isn't necessary.

* refactor(web): updating custom exception message creation

Uses new kork methods to clean up some error handling and custom exception message creation.

---------

Co-authored-by: David Byron <dbyron@salesforce.com>
Co-authored-by: Richard Timpson <richard.timpson@salesforce.com>