feat(pipeline): add a pipelineNameFilter query param to the /pipelines/{application} endpoint (#1504)
* feat(pipelineController): add a pipelineNameFilter query param to the /pipelines/{application} endpoint

This adds a pipelineNameFilter query parameter to the /pipelines/{application} endpoint. If pipelineNameFilter is present, the endpoint will return a list of pipelines whose pipeline name contains the pipelineNameFilter.

This change is necessary to enable some optimizations in the front end - namely, filtering pipelines on the backend, instead of always querying for every pipeline in an application, and filtering the list in the front end.

* feat(pipeline): check for null pipeline name and log error

* fix(test/pipeline): updating tests and fixing getPipelinesByApplication implementation.

* fix(pipeline): make the comparison case insensitive.

* refactor(tests): convert groovy tests to java

* fix(tests): address minor issues

---------

Co-authored-by: Richard Timpson <richard.timpson@salesforce.com>