perf(repository): Remove duplicate storeStage call (#4795)
* test(handler): Verify that start time is set

Verify that the start time for a stage is set even when we
encounter an exception

* perf(stage): Remove duplicate storeStage call

The duplicate repository.storeStage call was originally added in
https://github.com/spinnaker/orca/commit/e872ce859e5e82ca9a4726bea596ae1513b0e765
to ensure that the start time for a stage is set even if the
handler encounters an exception. However, even without the extra
repository.storeStage, orca still sets the start time for a stage
when it encounters an exception.

---------

Co-authored-by: Daniel Zheng <d.zheng@salesforce.com>