refactor(spring-security): refactor spring security and spring from 5.x to 6.x with spring boot upgrade to 3.x
With spring boot upgrade, spring security also upgrades from 5.x to 6.x. As per the migration [steps](https://www.baeldung.com/spring-security-migrate-5-to-6), `WebSecurityConfigurerAdapter` has been removed. So, it is not required to be extended, instead bean can be registered.

`WebMvcConfigurerAdapter` class has been deprecated in Spring 5.0 and subsequently removed from spring 6.0. So, replacing it with `WebMvcConfigurer` interface for direct implementation.
https://docs.spring.io/spring-framework/docs/5.0.0.RELEASE/javadoc-api/org/springframework/web/servlet/config/annotation/WebMvcConfigurerAdapter.html

Removing bean of WebSecurityConfigurerAdapter, as in spring security 6.0 it is no more required.
https://github.com/spring-projects/spring-security/pull/11923
https://www.baeldung.com/spring-security-migrate-5-to-6#2-websecurityconfigureradapter